Replace dynamicRegistration with persistRegistry, rename url to namespace
