+++ src/middleware/packages/ontologies/actions/getRdfPrefixes.js
@@ -5,7 +5,7 @@ module.exports = {
     const ontologies = await this.actions.list({}, { parentCtx: ctx });
     return ontologies
       .sort((a, b) => (a.prefix < b.prefix ? -1 : a.prefix > b.prefix ? 1 : 0))
-      .map(ontology => `PREFIX ${ontology.prefix}: <${ontology.url}>`)
+      .map(ontology => `PREFIX ${ontology.prefix}: <${ontology.namespace}>`)
       .join('\n');
   }
 };
