+++ src/middleware/packages/ontologies/actions/list.js
@@ -2,10 +2,10 @@ module.exports = {
   visibility: 'public',
   cache: true,
   async handler(ctx) {
-    if (this.settings.dynamicRegistration) {
+    if (this.settings.persistRegistry) {
       return await ctx.call('ontologies.registry.list');
     } else {
-      return this.settings.ontologies;
+      return Object.values(this.ontologies);
     }
   }
 };
