+++ src/middleware/packages/ontologies/ontologies/core/acl.json
@@ -1,4 +1,4 @@
 {
   "prefix": "acl",
-  "url": "http://www.w3.org/ns/auth/acl#"
+  "namespace": "http://www.w3.org/ns/auth/acl#"
 }
