+++ src/middleware/packages/ontologies/ontologies/core/foaf.json
@@ -1,4 +1,4 @@
 {
   "prefix": "foaf",
-  "url": "http://xmlns.com/foaf/0.1/"
+  "namespace": "http://xmlns.com/foaf/0.1/"
 }
