+++ src/middleware/packages/ontologies/ontologies/core/rdf.json
@@ -1,4 +1,4 @@
 {
   "prefix": "rdf",
-  "url": "http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+  "namespace": "http://www.w3.org/1999/02/22-rdf-syntax-ns#"
 }
