+++ src/middleware/packages/ontologies/ontologies/core/rdfs.json
@@ -1,4 +1,4 @@
 {
   "prefix": "rdfs",
-  "url": "http://www.w3.org/2000/01/rdf-schema#"
+  "namespace": "http://www.w3.org/2000/01/rdf-schema#"
 }
