+++ src/middleware/packages/ontologies/ontologies/core/xsd.json
@@ -1,4 +1,4 @@
 {
   "prefix": "xsd",
-  "url": "http://www.w3.org/2001/XMLSchema#"
+  "namespace": "http://www.w3.org/2001/XMLSchema#"
 }
