+++ src/middleware/packages/ontologies/ontologies/custom/mp.json
@@ -1,4 +1,4 @@
 {
   "prefix": "mp",
-  "url": "http://virtual-assembly.org/ontologies/pair-mp#"
+  "namespace": "http://virtual-assembly.org/ontologies/pair-mp#"
 }
