+++ src/middleware/packages/ontologies/ontologies/custom/oasis.json
@@ -1,4 +1,4 @@
 {
   "prefix": "oasis",
-  "url": "http://cooperative-oasis.org/ns/core#"
+  "namespace": "http://cooperative-oasis.org/ns/core#"
 }
