+++ src/middleware/packages/ontologies/ontologies/custom/og.json
@@ -1,6 +1,6 @@
 {
   "prefix": "og",
-  "url": "http://virtual-assembly.org/ontologies/organigraph#",
+  "namespace": "http://virtual-assembly.org/ontologies/organigraph#",
   "owl": "https://raw.githubusercontent.com/assemblee-virtuelle/organigraph/master/ontology/organigraph.owl",
   "jsonldContext": {
     "og:leads": {
