+++ src/middleware/tests/ontologies/ontologies/ont1.json
@@ -1,6 +1,6 @@
 {
   "prefix": "ont1",
-  "url": "https://www.w3.org/ns/ontology1#",
+  "namespace": "https://www.w3.org/ns/ontology1#",
   "jsonldContext": "https://www.w3.org/ns/ontology1.jsonld",
   "preserveContextUri": true
 }
