+++ src/middleware/tests/ontologies/ontologies/ont2.json
@@ -1,5 +1,5 @@
 {
   "prefix": "ont2",
-  "url": "https://www.w3.org/ns/ontology2#",
+  "namespace": "https://www.w3.org/ns/ontology2#",
   "jsonldContext": "https://www.w3.org/ns/ontology2.jsonld"
 }
