+++ src/middleware/tests/ontologies/ontologies/ont3.json
@@ -1,6 +1,6 @@
 {
   "prefix": "ont3",
-  "url": "https://www.w3.org/ns/ontology3#",
+  "namespace": "https://www.w3.org/ns/ontology3#",
   "jsonldContext": {
     "friend": {
       "@id": "ont3:friend",
